Complete image conversion revamp
All image conversion logic is now performed in the PHNImage class, which
can be used outside of widgets. The image viewer makes use of this image
class to display the contents, where the PHNImage class prepares a
pixmap for display purposes. Whenever changes are made to the image, the
changed signal is fired allowing viewer and others to refresh. The
conversion logic has been slightly adjusted to fix bugs and improve
performance. More changes may follow.