Hidden first 32 RISC registers, added value logic
The first 32 registers are restricted to the executing RISC processor
logic. Writing to it causes unexpected behavior, most often it will
simply crash. There is no purpose for them, so they are hidden using an
address offset boundary. Finally, the value text changes by the user can
be used to adjust the register value using hex or decimal.