Proper support for service mode in the STK500 logic
Right after reset, service mode is detected by sending and receiving a
token String. When inside service mode, standard STK commands are
disabled and return an error. The task handler allows tasks that don't
require the firmware to run to execute still. As well, a proper status
is included.

All of this allows you to recover firmware onto the device in the event
of an upload failure.