Implement dynamic firmware UART baud + multiserial
By writing to the UART registers it is possible to change the baud rate
of the serial UART hardware components. Included is a timing-critical
implementation to switch the baud rate of firmware operation without
locking out. This, needed for multi-serial later on, which requires
operation at a variety of baud rates.