Simplify overall protocol (status) handling
The port is now stored within the stk500 protocol class. In the future
all port functions should be inside the stk500 class so there is no need
for an exposed stk500Port at all.

The status reporting is improved by using the protocol directly, instead
of having a separate parameter for updateStatus. This means that an
updated status will almost always reflect the current state of the
stk500 protocol. To properly show an unopened state, UNOPENED was added,
which is also used to close the port.