Add firmware update functionality for old versions
If current firmware is outdated, the toolkit uploads a sketch to the
device that, through a hack of calling instructions in the old firmware,
is capable of flashing the first two pages of the firmware into locked
service mode. The device can then be updated through the usual service
mode.