Simplified service routine programming using sketch
Previously the first page of firmware flash was used solely to program
the firmware. However, as demonstrated with the recovery sketch, this is
not at all required. The service routine now sits on a sketch and makes
use of a single function in firmware (spm) to program the device. The
service programming protocol is extended to deploy the SPM function on
the first page while flashing. It jumps to sketch area if flashing is
abrupted, executing the service routine, so no harm can be done. In
fact, security is improved, by occupying only the first page and not the
first two pages while flashing. Most importantly, it allows the firmware
to be reduced in size by over 256 bytes.